+++ packages/accounts/accounts-tab/admin.js
@@ -75,10 +75,11 @@ var tabularOptions = {
       }
     },
     {
-      data: 'roles()',
+      data: 'roles',
       title: orion.helpers.getTranslation('accounts.index.tableTitles.roles'),
       render: function(val, type, doc) {
-        return val.map(function(role) {
+        var roles = Roles._collection ? doc.roles() : val || [];
+        return roles.map(function(role) {
           return '<span class="label label-danger red">' + role + '</span>';
         }).join('');
       }