+++ packages/relationships/attribute.js
@@ -11,11 +11,22 @@ var getSchema = function(options, hasMany) {
     create: Match.Optional(Function),
     additionalFields: Match.Optional(Array),
     sortFields: Match.Optional(Match.OneOf(Array, Object)),
+    validateOnClient: Match.Optional(Boolean),
+    validateOnServer: Match.Optional(Boolean),
+    dontValidate: Match.Optional(Boolean),
     render: Match.Optional({
       item: Function,
       option: Function
     })
   }));
+
+  if (!_.has(options, 'validateOnClient')) {
+    options.validateOnClient = true;
+  }
+
+  if (!_.has(options, 'validateOnServer')) {
+    options.validateOnServer = true;
+  }
 
   if (!options.filter) {
     options.filter = function(userId) {
@@ -44,12 +55,8 @@ var getSchema = function(options, hasMany) {
 
   if (!options.render) {
     options.render = {
-      item: function(item, escape) {
-        return render_item_default(item, escape);
-      },
-      option: function(item, escape) {
-        return render_item_default(item, escape);
-      }
+      item: render_item_default,
+      option: render_item_default
     };
   }
 
@@ -90,12 +97,28 @@ var getSchema = function(options, hasMany) {
     }
   }
 
-  if (hasMany) {
+  if (options.dontValidate && hasMany) {
+    return {
+      type: [String],
+      orion: options
+    };
+  } else if (options.dontValidate && !hasMany) {
+    return {
+      type: String,
+      orion: options,
+    };
+  } else if (hasMany) {
     return {
       type: [String],
       orion: options,
       custom: function() {
-        if (this.isSet && _.isArray(this.value)) {
+        if (Meteor.isClient && !options.validateOnClient) {
+          return;
+        }
+        if (Meteor.isServer && !options.validateOnServer) {
+          return;
+        }
+        if (this.isSet && _.isArray(this.value) && this.value) {
           var count = options.collection.find({ $and: [{ _id: { $in: this.value } }, options.filter(this.userId)] }).count();
           if (count != this.value.length) {
             return 'notAllowed';
@@ -108,7 +131,13 @@ var getSchema = function(options, hasMany) {
       type: String,
       orion: options,
       custom: function() {
-        if (this.isSet && _.isString(this.value)) {
+        if (Meteor.isClient && !options.validateOnClient) {
+          return;
+        }
+        if (Meteor.isServer && !options.validateOnServer) {
+          return;
+        }
+        if (this.isSet && _.isString(this.value) && this.value) {
           var count = options.collection.find({ $and: [{ _id: this.value }, options.filter(this.userId)] }).count();
           if (count != 1) {
             return 'notAllowed';